requirements.txt
requests
pillow
pillow_heif
ollama
tqdm
